pyproject.toml
[project]
name = "plant-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flax>=0.12.1",
    "google-metrax>=0.2.4",
    "matplotlib>=3.10.7",
    "polars>=1.35.2",
    "tensorflow>=2.20.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
